+++ code/game/objects/items/apc_frame.dm
@@ -10,6 +10,8 @@
 /obj/item/frame/apc/attackby(obj/item/weapon/W as obj, mob/user as mob)
 	..()
 	if (iswrench(W))
+		if (!W.tool_is_usable())
+			return
 		new /obj/item/stack/material/steel( get_turf(src.loc), 2 )
 		qdel(src)
